+++ src/htif-factory.cpp
@@ -29,7 +29,7 @@ static bool htif_peek(const pma_entry &pma, const machine &m, uint64_t page_offs
     return (page_offset % PMA_PAGE_SIZE) == 0 && page_offset < pma.get_length();
 }
 
-pma_entry make_htif_pma_entry(uint64_t start, uint64_t length) {
+pma_entry make_htif_pma_entry(uint64_t start, uint64_t length, htif_runtime_config *context) {
     pma_entry::flags f{
         true,                // R
         true,                // W
@@ -38,7 +38,7 @@ pma_entry make_htif_pma_entry(uint64_t start, uint64_t length) {
         false,               // IW
         PMA_ISTART_DID::HTIF // DID
     };
-    return make_device_pma_entry("HTIF device", start, length, htif_peek, &htif_driver).set_flags(f);
+    return make_device_pma_entry("HTIF device", start, length, htif_peek, &htif_driver, context).set_flags(f);
 }
 
 } // namespace cartesi