+++ src/htif.cpp
@@ -16,6 +16,7 @@
 
 #include "htif.h"
 #include "i-device-state-access.h"
+#include "machine-runtime-config.h"
 #include "pma-constants.h"
 #include "strict-aliasing.h"
 #include "tty.h"
@@ -90,12 +91,17 @@ static execute_status htif_yield(i_device_state_access *a, uint64_t cmd, uint64_
     return status;
 }
 
-static execute_status htif_console(i_device_state_access *a, uint64_t cmd, uint64_t data) {
+static execute_status htif_console(htif_runtime_config *runtime_config, i_device_state_access *a, uint64_t cmd,
+    uint64_t data) {
     // If console command is enabled, perform it and acknowledge
     if (cmd < 64 && (a->read_htif_iconsole() >> cmd) & 1) {
         if (cmd == HTIF_CONSOLE_PUTCHAR) {
             uint8_t ch = data & 0xff;
-            tty_putchar(ch);
+            // In microarchitecture runtime_config will always be nullptr,
+            // therefore the HTIF runtime config is actually ignored.
+            if (!runtime_config || !runtime_config->no_console_putchar) {
+                tty_putchar(ch);
+            }
             a->write_htif_fromhost(HTIF_BUILD(HTIF_DEVICE_CONSOLE, cmd, 0));
         } else if (cmd == HTIF_CONSOLE_GETCHAR) {
             // In blockchain, this command will never be enabled as there is no way to input the same character
@@ -110,7 +116,8 @@ static execute_status htif_console(i_device_state_access *a, uint64_t cmd, uint6
     return execute_status::success;
 }
 
-static execute_status htif_write_tohost(i_device_state_access *a, uint64_t tohost) {
+static execute_status htif_write_tohost(htif_runtime_config *runtime_config, i_device_state_access *a,
+    uint64_t tohost) {
     // Decode tohost
     uint32_t device = HTIF_DEV_FIELD(tohost);
     uint32_t cmd = HTIF_CMD_FIELD(tohost);
@@ -122,7 +129,7 @@ static execute_status htif_write_tohost(i_device_state_access *a, uint64_t tohos
         case HTIF_DEVICE_HALT:
             return htif_halt(a, cmd, data);
         case HTIF_DEVICE_CONSOLE:
-            return htif_console(a, cmd, data);
+            return htif_console(runtime_config, a, cmd, data);
         case HTIF_DEVICE_YIELD:
             return htif_yield(a, cmd, data);
         //??D Unknown HTIF devices are silently ignored
@@ -134,7 +141,8 @@ static execute_status htif_write_tohost(i_device_state_access *a, uint64_t tohos
 /// \brief HTIF device write callback. See ::pma_write.
 static execute_status htif_write(void *context, i_device_state_access *a, uint64_t offset, uint64_t val,
     int log2_size) {
-    (void) context;
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    htif_runtime_config *runtime_config = reinterpret_cast<htif_runtime_config *>(context);
     // Our HTIF only supports 64-bit writes
     if (log2_size != 3) {
         return execute_status::failure;
@@ -143,7 +151,7 @@ static execute_status htif_write(void *context, i_device_state_access *a, uint64
     // Only these 64-bit aligned offsets are valid
     switch (offset) {
         case htif_tohost_rel_addr:
-            return htif_write_tohost(a, val);
+            return htif_write_tohost(runtime_config, a, val);
         case htif_fromhost_rel_addr:
             a->write_htif_fromhost(val);
             return execute_status::success;