+++ src/machine-c-api.h
@@ -357,9 +357,15 @@ typedef struct { // NOLINT(modernize-use-using)
     uint64_t update_merkle_tree;
 } cm_concurrency_config;
 
+/// \brief HTIF runtime configuration
+typedef struct { // NOLINT(modernize-use-using)
+    bool no_console_putchar;
+} cm_htif_runtime_config;
+
 /// \brief Machine runtime configuration
 typedef struct { // NOLINT(modernize-use-using)
     cm_concurrency_config concurrency;
+    cm_htif_runtime_config htif;
     bool skip_root_hash_check;
     bool skip_version_check;
 } cm_machine_runtime_config;